+++ caddyconfig/httpcaddyfile/blocktypes/globalblock/globalblock.go
@@ -19,9 +19,9 @@ import (
 
 	"github.com/caddyserver/caddy/v2/caddyconfig"
 	"github.com/caddyserver/caddy/v2/caddyconfig/caddyfile"
+	"github.com/caddyserver/caddy/v2/caddyconfig/configbuilder"
 	"github.com/caddyserver/caddy/v2/caddyconfig/httpcaddyfile"
-	"github.com/caddyserver/caddy/v2/caddyconfig/xcaddyfile/blocktypes"
-	"github.com/caddyserver/caddy/v2/caddyconfig/xcaddyfile/configbuilder"
+	"github.com/caddyserver/caddy/v2/caddyconfig/httpcaddyfile/blocktypes"
 )
 
 func init() {
