--- caddyconfig/xcaddyfile/example.xcaddyfile
@@ -1,41 +1,0 @@
-# XCaddyfile Example
-#
-# XCaddyfile uses explicit block type declarations with [type] syntax
-# to support multiple protocol types in a single configuration file.
-
-# [global] block configures both core Caddy settings and HTTP app options
-# This maintains backwards compatibility with standard Caddyfile global options
-[global] {
-	admin unix//run/caddy/admin.socket
-	log DEBUG
-	grace_period 30s
-	https_port 8443
-	http_port 8080
-}
-
-# [http.server] blocks configure individual HTTP sites
-# These are child blocks of [global] and add servers to the HTTP app
-[http.server] https://example.com {
-	root * /var/www/example
-	file_server
-	encode gzip
-}
-
-[http.server] https://api.example.com {
-	reverse_proxy localhost:9000
-}
-
-# Future extensions could include other protocol types:
-# [layer4] 0.0.0.0:22 {
-#     @ssh ssh
-#     route @ssh {
-#         proxy 127.0.0.1:1234
-#     }
-# }
-#
-# [pggat] 0.0.0.0:5432 {
-#     ssl self_signed
-#     pool default {
-#         address localhost:15432
-#     }
-# }
